acceleration to float conversion (not tested!)

+++ Src/acceleration_sensor.c
@@ -82,7 +82,7 @@ HAL_StatusTypeDef ACC_getAllValues(I2C_HandleTypeDef *i2cHandler, s_acceleromete
 	
 	status = i2c_read_register(i2cHandler, ACCELEROMETER_I2C_ADRESS, ACC_OUT_X_MSB, (uint8_t*)&receiveBuffer, 6);
 	
-	if (status!=HAL_OK)return status; // bei Fehler Funktion hier mit status Rueckgabe verlassen
+	if (status!=HAL_OK) return status; // bei Fehler Funktion hier mit status Rueckgabe verlassen
 	//X_Wert_14_0 = (X_MSB<<5) || (X_LSB>>2)
 	//bei Fehlerhaften Werten, evtl vor "<<" typecast auf int16 
 	accValues->x_Value = 0;
@@ -109,7 +109,7 @@ HAL_StatusTypeDef ACC_getAllValues(I2C_HandleTypeDef *i2cHandler, s_acceleromete
 // Convert Received Data in Acceleration float
 // Pram: breiteInBit ist entweder 8 oder 12bit / messbereich +-2, 4, 8 g
 // für das speichern der Werte wäre uint16_t sinnvoller als float
-float ACC_convertAccelToFloat(int16_t rohDaten, uint8_t breiteInBit, uint8_t messbereich)
+float ACC_convertAccelToFloat(int16_t rohDaten, uint8_t breiteInBit, float messbereich)
 {
 	// 0111 1111 1111 = +1.999 / 3.998 / 7.996
 	// 0000 0000 0001 = 0.001 / 0.002 / 0.004
@@ -117,7 +117,7 @@ float ACC_convertAccelToFloat(int16_t rohDaten, uint8_t breiteInBit, uint8_t mes
 	// 1000 0000 0000 = -2 / -4 / -8
 	float accelerationFloat;
 	
-	accelerationFloat = (rohDaten * messbereich) / (float)(1<<(breiteInBit-1)); //TODO?
+	accelerationFloat = (rohDaten * messbereich) / (1<<(breiteInBit-1)); //TODO?
 	
 	return accelerationFloat;
 }
